requirements.txt
Django==2.0.6
Pillow==5.1.0
PyJWT==1.6.4
Unidecode==1.0.22
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.7.45
botocore==1.10.45
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
django-facebook==6.0.3
django-filter==1.1.0
django-forms-bootstrap==3.1.0
django-storages==1.6.6
docutils==0.14
facebook-sdk==2.0.0
gunicorn==19.8.1
idna==2.7
jmespath==0.9.3
lxml==4.2.2
oauthlib==2.1.0
psycopg2==2.7.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
social-auth-app-django==2.1.0
social-auth-core==1.7.0
twython==3.7.0
urllib3==1.23
